requirements.txt
fastapi
uvicorn
moviepy==1.0.3
librosa
numpy
requests
pillow
opencv-python